feat: refactor class and detail class pages, implement assignment card component
